requirements.txt
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
requests>=2.31.0
cryptography>=42.0.0
python-dotenv>=1.0.0
schedule>=1.2.1
psycopg2-binary>=2.9.0